check if file exists

+++ build-package.ps1
@@ -3,8 +3,12 @@ $destination = "HttpKeyboardMouse\bin\Release\net6.0-windows"
 $packageName = "package.zip"
 
 # don't dot net clean, remove the whole directory in case things have been renamed
-Remove-Item -Recurse -Force $destination
-Remove-Item -Force $packageName
+if (Test-Path $destination ) {
+  Remove-Item -Recurse -Force $destination 
+}
+if (Test-Path $destination ) {
+    Remove-Item -Force $packageName
+}
 
 dotnet build --configuration Release
